Move DNA accessibility component files to new directory
