+++ src/components/dna-accessibility/DnaAccessibilitySettings.js
@@ -9,9 +9,9 @@ import FormLabel from '@material-ui/core/FormLabel';
 import RadioGroup from '@material-ui/core/RadioGroup';
 import { makeStyles } from '@material-ui/core/styles';
 
-import AntSwitch from './AntSwitch';
+import AntSwitch from '../../AntSwitch';
 
-import { useDnaAccessLabelStyle, useDnaAccessRowNorm } from './state';
+import { useDnaAccessLabelStyle, useDnaAccessRowNorm } from '../../state';
 
 const useStyles = makeStyles((theme) => ({
   iconRadioLegend: {
